requirements.txt
pandas
numpy
requests
yfinance
beautifulsoup4
selenium
webdriver-manager
scikit-learn
torch
matplotlib
seaborn
python-dotenv
feedparser
